+++ scratch.js
@@ -1,23 +1,28 @@
 
 import nlp from './src/three.js'
-// import plg from './plugins/dates/src/plugin.js'
+import plg from './plugins/dates/src/plugin.js'
 // import plg from './plugins/_experiments/cmd-k/src/plugin.js'
 // nlp.plugin(plg)
 
 
 // already tested for
-nlp('simon says run').debug()
+// nlp('simon says run').debug()
 
 // redact misses google
-console.log(nlp('Mary joined Google today').redact().text())
+// console.log(nlp('Mary joined Google today').debug().redact().text())
+
+
+nlp('I live at 4 main st.').debug()
+// nlp('Call John Smith at (800) 555-0000.').debug().redact().debug()
+// nlp('i met her').debug()
 
 // paris has trailing period
-console.log(nlp('I went to Paris.').places().out('array'))
+// console.log(nlp('I went to Paris.').places().out('array'))
 
 // missing methods
-nlp('it cost $2.50').money().currency()   // ❌ TypeError
-nlp('two thirds').fractions().toText()    // ❌ TypeError
-nlp('80%').percentages().toFraction()     // ❌ TypeError
-nlp('$4.09CAD').money().currency()
-nlp('1/2').fractions().toText()
-nlp('80%').percentages().toFraction()
+// nlp('it cost $2.50').money().currency()   // ❌ TypeError
+// nlp('two thirds').fractions().toText()    // ❌ TypeError
+// nlp('80%').percentages().toFraction()     // ❌ TypeError
+// nlp('$4.09CAD').money().currency()
+// nlp('1/2').fractions().toText()
+// nlp('80%').percentages().toFraction()